# apps/api/openapi.yaml
openapi: 3.0.3

info:
  title: Task API
  version: 1.0.0
  description: Basic task tracking API for CS453 project


servers:
  - url: http://localhost:3000


paths:

  /health:
    get:
      summary: check API /health
      responses: 
        "200":
          description: Server is running
          content:
            application/json:
              schema: 
                type: object
                properties:
                  status: 
                    type: string
                    example: ok
                  
  /tasks:

    get:
      summary: Get all tasks
      responses:

        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"



    post:
      summary: Create a task

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:

                title:
                  type: string
                  example: Create task API

                status:
                  type: string
                  example: todo


      responses:

        "201":
          description: Task created

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"


        "400":
          description: Missing title
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"


  /tasks/{id}:
    get:
      summary: Get a task by id
      description: Returns a task
      parameters:
        - name: id
          in: path
          required: true
          description: Task id
          schema:
            type: integer
       
      responses: 
        "200":
          description: Task found
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

        "404":
          description: Task not found
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "400":
          description: Invalid task id
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update a task
      description: Updates field/s of a task

      parameters:
        - name: id
          in: path
          required: true
          description: Task id
          schema: 
            type: integer

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTask"

      responses:
        "200": 
          description: Task updates succesfully
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Task"

        "400":
          description: invalid task id or missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
           

        "404":
          description: Task not found
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"


    delete: 
      summary: Delete a task
      description: Deletes a task

      parameters:
        - name: id 
          in: path
          required: true
          description: Task id 
          schema:
            type: integer

      responses:
        "204":
          description: Task deleted succesfully

        "400":
          description: Invalid task id 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"




components:

  schemas:

    Task:

      type: object
      required:
        - id
        - title
        - status

      properties:

        id:
          type: integer
          example: 1

        title:
          type: string
          example: Create task API

        status:
          type: string
          example: todo
    
    UpdateTask:
      type: object

      properties:
        title:
          type: string
          example: Update API documentation

        status:
          type: string
          example: done 


    Error:
      type: object

      properties:
        error:
          type: string
          example: Task not found